+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Allow users to manage the downloading of WASM grammars not bundled with treetags
 - Add native Terraform tag generator that uses downloadable WASM tree-sitter grammar
 - Add basic tags support for Zig
 - Add basic tags support for Terraform
+- Add basic tags support for Swift
 
 ### Changed
 - Tag Address is truncated to 96 chars for native parsers